+++ packages/output/src/common/output-channel.ts
@@ -15,19 +15,55 @@
  ********************************************************************************/
 
 import { Emitter, Event } from '@theia/core';
-import { injectable, inject } from 'inversify';
+import { injectable, inject, postConstruct } from 'inversify';
 import { OutputPreferences } from './output-preferences';
+import { Disposable, DisposableCollection } from 'vscode-ws-jsonrpc';
 
 @injectable()
-export class OutputChannelManager {
+export class OutputChannelManager implements Disposable {
     protected readonly channels = new Map<string, OutputChannel>();
+    protected selectedChannelValue: OutputChannel | undefined;
 
-    private readonly channelDeleteEmitter = new Emitter<{channelName: string}>();
-    private readonly channelAddedEmitter = new Emitter<OutputChannel>();
+    protected readonly channelDeleteEmitter = new Emitter<{ channelName: string }>();
+    protected readonly channelAddedEmitter = new Emitter<OutputChannel>();
+    protected readonly selectedChannelEmitter: Emitter<void> = new Emitter<void>();
+    protected readonly listOrSelectionEmitter: Emitter<void> = new Emitter<void>();
     readonly onChannelDelete = this.channelDeleteEmitter.event;
     readonly onChannelAdded = this.channelAddedEmitter.event;
+    readonly onSelectedChannelChange = this.selectedChannelEmitter.event;
+    readonly onListOrSelectionChange = this.listOrSelectionEmitter.event;
 
-    constructor(@inject(OutputPreferences) protected preferences: OutputPreferences) {
+    protected toDispose = new DisposableCollection();
+
+    @inject(OutputPreferences)
+    protected readonly preferences: OutputPreferences;
+
+    @postConstruct()
+    protected init(): void {
+        this.getChannels().forEach(this.registerListener.bind(this));
+        this.toDispose.push(this.onChannelAdded(channel => {
+            this.listOrSelectionEmitter.fire(undefined);
+            this.registerListener(channel);
+        }));
+        this.toDispose.push(this.onChannelDelete(event => {
+            this.listOrSelectionEmitter.fire(undefined);
+            if (this.selectedChannel && this.selectedChannel.name === event.channelName) {
+                this.selectedChannel = this.getVisibleChannels()[0];
+            }
+        }));
+    }
+
+    protected registerListener(outputChannel: OutputChannel): void {
+        if (!this.selectedChannel) {
+            this.selectedChannel = outputChannel;
+        }
+        this.toDispose.push(outputChannel.onVisibilityChange(event => {
+            if (event.visible) {
+                this.selectedChannel = outputChannel;
+            } else if (outputChannel === this.selectedChannel) {
+                this.selectedChannel = this.getVisibleChannels()[0];
+            }
+        }));
     }
 
     getChannel(name: string): OutputChannel {
@@ -43,23 +79,41 @@ export class OutputChannelManager {
 
     deleteChannel(name: string): void {
         this.channels.delete(name);
-        this.channelDeleteEmitter.fire({channelName: name});
+        this.channelDeleteEmitter.fire({ channelName: name });
     }
 
     getChannels(): OutputChannel[] {
         return Array.from(this.channels.values());
     }
+
+    getVisibleChannels(): OutputChannel[] {
+        return this.getChannels().filter(channel => channel.isVisible);
+    }
+
+    public dispose() {
+        this.toDispose.dispose();
+    }
+
+    get selectedChannel(): OutputChannel | undefined {
+        return this.selectedChannelValue;
+    }
+
+    set selectedChannel(channel: OutputChannel | undefined) {
+        this.selectedChannelValue = channel;
+        this.selectedChannelEmitter.fire(undefined);
+        this.listOrSelectionEmitter.fire(undefined);
+    }
 }
 
 export class OutputChannel {
 
-    private readonly visibilityChangeEmitter = new Emitter<{visible: boolean}>();
+    private readonly visibilityChangeEmitter = new Emitter<{ visible: boolean }>();
     private readonly contentChangeEmitter = new Emitter<OutputChannel>();
     private lines: string[] = [];
     private currentLine: string | undefined;
     private visible: boolean = true;
 
-    readonly onVisibilityChange: Event<{visible: boolean}> = this.visibilityChangeEmitter.event;
+    readonly onVisibilityChange: Event<{ visible: boolean }> = this.visibilityChangeEmitter.event;
     readonly onContentChange: Event<OutputChannel> = this.contentChangeEmitter.event;
 
     constructor(readonly name: string, readonly preferences: OutputPreferences) { }
@@ -95,7 +149,7 @@ export class OutputChannel {
 
     setVisibility(visible: boolean): void {
         this.visible = visible;
-        this.visibilityChangeEmitter.fire({visible});
+        this.visibilityChangeEmitter.fire({ visible });
     }
 
     getLines(): string[] {